create basic logic for logini window + refactor code for modularity
