+++ src/App.js
@@ -1,34 +1,28 @@
 import './App.css';
-import { useState, useEffect } from 'react';
-import { fetchAllBuildings, fetchBuildingById } from './services/buildingService';
-import { BuildingInfo } from './components/BuildingInfo';
-import { Select } from './components/Select';
+import { LoginPage } from './pages/LoginPage';
+import { BuildingInfoPage } from './pages/BuildingInfoPage';
+import { useState } from 'react';
 
 
 export default function App() {
-  const [allBuildings, setAllBuildings] = useState([]);
-  const [selectedBuilding, setselectedBuilding] = useState([]);
+const [isLoggedIn, setIsLoggedIn] = useState(false);
 
-  useEffect(() => {
-    fetchAllBuildings()
-    .then((data) => {
-      setAllBuildings(data);
-      if (data.length > 0)
-        setselectedBuilding([data[0]])
-    })
-  }, []);
-
-  function handleSelect(e) {
-    fetchBuildingById(e.target.value)
-    .then((data) => {setselectedBuilding(data);
-    })
-  };
+function handleLogin() {
+  if (!isLoggedIn)
+    setIsLoggedIn(true);
+  else 
+    setIsLoggedIn(false);
+}
 
   return (
     <>
       <div>
-        <Select idSelect='Building' nameSelect='Building 'handleSelect={ handleSelect } allBuildings={ allBuildings }/> 
-        {selectedBuilding && <BuildingInfo selectedBuilding={ selectedBuilding }/>}
+      {isLoggedIn ? (
+          <BuildingInfoPage />
+        ) : (
+          <LoginPage handleLogin={handleLogin} />
+        )}
+
       </div>
     </>
   );
