+++ src/components/BuildingInfo/BuildingInfo.js
